Update how error messages are displayed to match new error array format
The way errors are stored in the session was changed to use a 1D array
to help simplify displaying errors on a view and to match how request
data is stored, keeping things more consistent. This commit introduces
using the ':' to split up an error group and field within error
directives.