Fix Email Verification Middleware not working when user is not authed
Internally when checking if a user has a verified email, the users ID
is requested but when a user is not logged in, false is returned which
is still deemed valid by is_null(). The solution is to change this
return value to null when requesting the user ID if one does not
exist due to not being logged in.