Load third party Composer packages into memory during server boot-up
This commit uses the generated Composer classmap to try and preload all
Composer vendor/ packages using require_once, meaning this work does not
have to be done during the request cycle. No classmap lookup or
filesystem check will need to be done during a request, it is done
during the server boot-up phase.